Replace gRPC with a TCP socket

+++ recognize/Observer/src/observer/AsyncInference/types.hpp
@@ -5,10 +5,10 @@
 
 namespace AsyncInference {
     struct SingleDetection {
-        int x;
-        int y;
-        int width;
-        int height;
+        float x;
+        float y;
+        float width;
+        float height;
         float confidence;
         std::string label;
     };
